feat(sharing): 添加Sharing API的数据仓库和接口
添加了`SharingRepository`，用于封装`SharingService`的所有网络请求。此仓库提供了以下功能：
- `createShare`：创建可公开访问的音乐或视频URL。
- `deleteShare`：删除已存在的分享。
- `getShares`：获取用户可管理的分享媒体信息。
- `updateShare`：更新分享的描述或过期时间。

同时，修复了`SharingService`中`updateShare`接口的`expires`参数类型，将其从`Int`更正为`Long`，以匹配其在`SharingRepository`中的使用。

+++ lib_open_subsonic_api/src/main/java/com/che2n3jigw/android/libs/opensubsonicapi/service/SharingService.kt
@@ -64,7 +64,7 @@ interface SharingService {
     suspend fun updateShare(
         @Query("id") id: String,
         @Query("description") description: String? = null,
-        @Query("expires") expires: Int? = null
+        @Query("expires") expires: Long? = null
     ): BaseResponse<BaseSubsonicResponse>
 
 }
